requirements.txt
openai>=1.30.0
pinecone[grpc]>=5.0.0
streamlit>=1.35.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
pandas>=2.2.0
tqdm>=4.66.0
tenacity>=8.3.0
